test: controller 분리에 따른 테스트 이동
